Tests/LibWeb/Text/expected/grid/content-alignment-overflow.txt

left: 10.0
top: 10.0
right: 30.0
bottom: 30.0
